Avance

+++ src/main/java/org/iesvdm/dao/ClienteDAO.java
@@ -10,7 +10,7 @@ public interface ClienteDAO {
 	public void create(Cliente cliente);
 	
 	public List<Cliente> getAll();
-	public Optional<Cliente>  find(int id);
+	public Optional<Cliente>  find(long id);
 	
 	public void update(Cliente cliente);
